requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
dbus-python==1.2.8
decorator==4.3.2
docutils==0.14
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
jeepney==0.4
Kivy==1.10.1
Kivy-Garden==0.1.4
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
pydbus==0.6.0
Pygments==2.3.1
pyotp==2.2.7
python-networkmanager==2.1
requests==2.21.0
SecretStorage==3.1.1
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
